+++ src/main/java/net/nemezanevem/gregtech/api/gui/widgets/PhantomSlotWidget.java
@@ -2,22 +2,16 @@ package net.nemezanevem.gregtech.api.gui.widgets;
 
 import com.google.common.collect.Lists;
 import mezz.jei.api.gui.handlers.IGhostIngredientHandler.Target;
-import net.minecraft.client.KeyMapping;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.MouseHandler;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.renderer.Rect2i;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.ClickType;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.client.event.InputEvent;
 import net.minecraftforge.items.IItemHandlerModifiable;
 import net.nemezanevem.gregtech.api.gui.ingredient.IGhostIngredientTarget;
 import net.nemezanevem.gregtech.api.util.SlotUtil;
 import net.nemezanevem.gregtech.client.util.MouseButtonHelper;
-import net.nemezanevem.gregtech.client.util.TooltipHelper;
-import org.lwjgl.glfw.GLFW;
 
 import javax.annotation.Nonnull;
 import java.util.Collections;
@@ -37,7 +31,7 @@ public class PhantomSlotWidget extends SlotWidget implements IGhostIngredientTar
     }
 
     @Override
-    public boolean mouseClicked(int mouseX, int mouseY, int button) {
+    public boolean mouseClicked(double mouseX, double mouseY, int button) {
         if (isMouseOverElement(mouseX, mouseY) && gui != null) {
             if (button == 1 && clearSlotOnRightClick && !slotReference.getItem().isEmpty()) {
                 slotReference.set(ItemStack.EMPTY);
@@ -52,7 +46,7 @@ public class PhantomSlotWidget extends SlotWidget implements IGhostIngredientTar
     }
 
     @Override
-    public boolean mouseDragged(int mouseX, int mouseY, int button, double dragX, double dragY) {
+    public boolean mouseDragged(double mouseX, double mouseY, int button, double dragX, double dragY) {
         if (isMouseOverElement(mouseX, mouseY) && gui != null) {
             ItemStack is = gui.player.containerMenu.getCarried().copy();
             is.setCount(1);
